backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
pydantic==2.10.6
pydantic-settings==2.8.1
email-validator==2.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
neo4j==5.28.1
SQLAlchemy[asyncio]==2.0.38
psycopg[binary]==3.2.5
asyncpg==0.30.0
alembic==1.14.1
celery==5.4.0
redis==5.2.1
requests==2.32.3
paramiko==3.5.1
netmiko==4.5.0
napalm==5.0.0
pysnmp==7.1.16
boto3==1.36.26
azure-mgmt-network==28.1.0
httpx==0.28.1
aiosqlite==0.21.0
google-generativeai>=0.8.0
pytest==8.3.4
pytest-asyncio==0.25.3
